requirements.txt
SQLAlchemy==1.4.29
suds-py3==1.4.5.0
tensorflow==2.7.0
nltk==3.6.7

